Fix `tryToRewritePositionalFilterSupport`
The problem was that one array of varargs was too long, and therefore
had a `null` value. This should never have been rewritten like this in
the translation to Scala anyway. Other uses of arrays in this file have
also been updated to pass the arguments directly, although they were not
incorrect.